Set up Badge for extension icon for users to know if site they are on has images for extension and moved Save button and features
Summary:
check if images on site load instead of having to click extension. Also set up temporary badge. Will need to send message from inject to background.js

Added badge to extension button that tells user if images on webpage for extension. If not, badge with 0 shown and nothing happens if user clicks. If so, '+' badge shows. Now need to get rid of the renderNoImagesPage() in App.js

Removed the renderNoImagesPoage(), but for some reason still needed to render the NUXButton in its place

Reviewers: gabriella, zaire, zackargyle

Reviewed By: zackargyle

Subscribers: tilde, zackargyle

Differential Revision: https://phabricator.pinadmin.com/D103733